Adding the Sender Portion
Just finally adding the Sender portion for the Wireless Temperature Project. This side basically has a temperature sensor which is running on the I2C library. The temperature data is sent over serially using an Arduino, then it is sent over wireless by a Bluetooth module, and then received by the Receiver portion of the Project.