Adding the Receiver Prortion
After the data is received from the Sender, the Receiver will have to parse the incoming data. Meaning the receiver is going to organize the data for the LCD display and output the actually data we want to see. At the same time, it will also calculate the data received from the Sender to Fahrenheit and  Celsius.